wip: implement simple service

+++ internal/players/dto/player.go
@@ -4,11 +4,11 @@ package dto
 type Player struct {
 	Id     string  `json:"id"`
 	Name   string  `json:"name"`
-	Rating float32 `json:"rating"`
+	Rating float64 `json:"rating"`
 }
 
 // PlayerRequest defines model for PlayerRequest.
 type PlayerRequest struct {
 	Name   string  `json:"name"  binding:"required"`
-	Rating float32 `json:"rating" binding:"required,gte=0,lte=10"`
+	Rating float64 `json:"rating" binding:"required,gte=0,lte=10"`
 }
